refactor: reorganize util functions from lib/ to util/
汎用ユーティリティをlib/からutil/に移動し、ドメイン特化コードと汎用コードの責務を明確化しました。

Changes:
- Create src/util/ directory for generic utilities
- Move array.ts, date.ts, highlight.ts, query.ts to src/util/
- Update import paths in storage.ts and HistoryItem component
- All tests passing after refactoring
