+++ src/util/array.spec.ts
