+++ src/util/array.ts
